Use 32 bit length fields for content instead of 64 bit.
The content is a single vector tile layer. If this is larger than 4
GBytes, we are doing something wrong.

+++ include/tgd_header/types.hpp
@@ -44,7 +44,7 @@ namespace tgd_header {
 
     using name_length_type = std::uint16_t;
 
-    using content_length_type = std::uint64_t;
+    using content_length_type = std::uint32_t;
 
 } // namespace tgd_header
